+++ tests/podmander-controller-strategies-first_available_tests.ads
@@ -1,0 +1,10 @@
+--  Copyright (C) 2026 Jochen Lillich
+--  SPDX-License-Identifier: Apache-2.0
+
+with AUnit.Test_Suites;
+
+package Podmander.Controller.Strategies.First_Available_Tests is
+
+   function Suite return AUnit.Test_Suites.Access_Test_Suite;
+
+end Podmander.Controller.Strategies.First_Available_Tests;